requirements.txt
streamlit
fastapi
uvicorn
python-multipart
pillow
face_recognition